+++ packages/nestjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chetodb/nestjs-business-central
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @chetodb/business-central@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes